Replace use of reflection for written book data
This makes use of new API since Bukkit 1.9.4, that gets and sets book
generation data more easily.

For 1.8.8 compatibility, a LegacyBookItem class has been added.
BookFactory determines which class to use by catching
NoClassDefFoundError.

See https://hub.spigotmc.org/stash/projects/SPIGOT/repos/bukkit/pull-requests/193/overview
